chore(parser): adjust authorship calculation in scoring
Refactored authorship percentage logic and integrated it into reward calculations.
