+++ src/parser/github-comment-module.ts
@@ -276,6 +276,7 @@ export class GithubCommentModule extends BaseModule {
         content: commentScore.score?.formatting,
         regex: commentScore.score?.words,
         ...(commentScore.score?.weight !== undefined && { weight: commentScore.score.weight }),
+        authorship: commentScore.score?.authorship,
       }).replace(/[\n\r]/g, "&#13;");
       // Makes sure any HTML injected in the templated is not rendered itself
       const sanitizedContent = commentScore.content
