+++ src/types/results.ts
@@ -80,5 +80,6 @@ export interface GithubCommentScore {
     priority?: number;
     reward: number;
     weight?: number;
+    authorship: number;
   };
 }
